logind: Remove session type check
c138da233b618ab5d42e4810f1ad7f2e687f2a59 introduced a few fallback
mechanisms for session lookup for situations where the current process
is not associated with an active session.

When using these fallbacks, the type of a found session would be checked
against a list of graphical environments, and session lookup would fail
if it did not match.

Remove this check as it serves no useful purpose.