backend/session: Run libseat callbacks on idle
libseat callbacks can be triggered from the stack of a libseat call,
which might come as a surprise. Postpone the callbacks until event loop
idle to break the stack.

It is possible to patch libseat to not execute events from within libset
calls, but that approach requires always calling libseat dispatch until
it returns 0 before sleeping the event loop, which adds permanent work.
This seems suboptimal as callbacks very rarely need to execute.